Add @reduxjs/toolkit After looking through redux documentation, this package makes supporting a redux based app much easier so I am implementing it into the project and will refactor app to accomodate.
