+++ src/components/common/DrinkCard/DrinkCard.js
@@ -43,7 +43,7 @@ function DrinkCard({drink, updateFavDrinks, username, favDrinks, isAuthenticated
   function hasUserFaved() {
     // must return negation because every only stops when evaled false
     // this keeps functionality while allowing logic to make sense
-    return !favDrinks.drinks.every((currDrink) => {
+    return !favDrinks.every((currDrink) => {
       return currDrink.idDrink !== drink.idDrink;
     });
   }
@@ -102,7 +102,7 @@ function DrinkCard({drink, updateFavDrinks, username, favDrinks, isAuthenticated
 function mapStateToProps(user) {
   return {
     username: user.username,
-    favDrinks: user.fav_drinks,
+    favDrinks: user.favDrinks,
     isAuthenticated: user.isAuthenticated
   }
 }
