+++ src/store/reducers/favDrinksReducer.js
@@ -1,0 +1,11 @@
+import { createSlice } from '@reduxjs/toolkit'
+
+const initialState = []
+
+const favDrinks = createSlice({
+  name: 'favDrinks',
+  initialState,
+  reducers: {}
+})
+
+export default favDrinks.reducer
